Removing Graph, Getting Form Interactivity
Since the ng2-charts was not playing nice with
my angular cli, I decided to ditch it. I don't need
a chart anyway, since I will be displaying information
with a table, just would've been nice to have is all.

The form for the user input is correctly working.
I haven't implemented a table or any calculations just yet,
all I am doing is displaying the APY on the frontend if the
user has pressed the calculate button. This is to check
that my form submission is working correctly in the first place

I will need to implement the ability to select either the month or the
year for additional contributions and APY. You should not be able to select
both, but currently you are able to. My implementation is probably off,
as I am using two radio inputs and not a connected, cohesive unit.
I'm sure Angular provides these options, so I am going to look into more right now
to see if that is a valid option.