Created InvestementTableCalc Service
I figured out the issue with the radio buttons:
the names were not the same, so they were able to be
active at the same time. HTML uses grouping via naming convention, so things
with similar names will not be able to hold an active state
at the same time

I created the table to output the information from the user.
This is a simple table that uses a for loop.
Angular 19 introduces the @for as a replacement for ngFor,
so I used that for this project to keep up to date on the latest
and greatest models

Created an interface for the table data, so we can pass around
objects that have that naming convention.

Implemented a Service for calculating the Investment Table.
While this could have been in a component file, I thought it best
to maintain the habit of registering services, since the Investment Calculator
theoretically could be used to find all sorts of things if this
is expanded upon in the future

However, now that the main functionality is implemented,
future commits will be used for styling. I will then look to refactor
code to get rid of redundancies and anything else repetitive
that could be its own code block

In mindfulness of CSS styling, it's possible that I will need to create
some shared components, as the table, user input, and header will
likely have similar backgrounds and functionalities. This will
make ng-content a useful feature